update parse post content
