+++ src/lib/util/parse-post-content.ts
@@ -1,61 +1,217 @@
-import { Agent } from "@/types/mod";
-import { recordMatch } from "@/lib/util/match";
+import type { Forward } from "@/lib/util/forwards";
+
 import { getDB } from "@/lib/db";
-import { replaceAndAccumulate } from "@/lib/util/replace-mentions";
-import replaceMentions from "@/lib/util/replace-mentions";
 import { TABLE_AGENT } from "@/config/mod";
 import { createTempAgent } from "@/lib/database/agent";
 import { StringRecordId } from "surrealdb.js";
-import { getAgent } from "@/lib/database/agent";
-
-const getAgentIdByNameOrCreate = async (
+import { docSort, head, MENTION_MATCH, merge, tail } from "@/lib/util/forwards";
+import replaceAll from "@/lib/util/replace-all";
+import { PROMPTS_SUMMARIZE } from "@/lib/templates/static";
+import { summarize } from "@/lib/ai";
+const getAgentIdFromName = async (
   name: string,
-  context: string = name,
-): Promise<Agent> => {
+  context?: string,
+): Promise<`agent:${string}`> => {
+  if (name.startsWith("agent:")) {
+    return confirmAgentId(name, context);
+  }
   const db = await getDB();
   try {
-    const [[agent]]: [[Agent]] = await db.query(
-      "SELECT * FROM type::table($table) WHERE name = $name",
+    const [[agent]]: [[{ id: `agent:${string}` }]] = await db.query(
+      `SELECT string::concat("", id) as id FROM type::table($table) WHERE name = $name`,
       {
         table: TABLE_AGENT,
         name,
       },
     );
-    return (agent ||
-      (await createTempAgent({ name, context }))) as Agent;
+    if (agent) {
+      return agent.id;
+    }
+    const temp = await createTempAgent({ name, context });
+    return temp.id.toString() as `agent:${string}`;
   } finally {
     await db.close();
   }
+  // return `agent:${name}`;
 };
-
-const replaceAgents = async (
-  name: string,
-  context: string = name,
-): Promise<string | [string, any]> => {
-  if (recordMatch.test(name)) {
-    if (name.startsWith("@agent:")) {
-      const agent = await getAgent(new StringRecordId(name.slice(1)));
-      return [name, agent];
+const confirmAgentId = async (
+  id: string,
+  context?: string,
+): Promise<`agent:${string}`> => {
+  if (!id.startsWith("agent:")) {
+    return getAgentIdFromName(id, context);
+  }
+  const rid = new StringRecordId(id);
+  const db = await getDB();
+  try {
+    const [[agent]]: [[{ id: `agent:${string}` }]] = await db.query(
+      `SELECT string::concat("", id) as id FROM type::table($table) WHERE id = $id`,
+      {
+        table: TABLE_AGENT,
+        id: rid,
+      },
+    );
+    if (agent) {
+      return agent.id;
     }
-    return name;
+    const [_, id_suffix] = id.split(":");
+    const name = await summarize(
+      id_suffix,
+      PROMPTS_SUMMARIZE.LLM_NAME_FROM_RANDOM,
+    );
+    const newAgent = await createTempAgent({ name, id_suffix, context });
+    return newAgent.id.toString() as `agent:${string}`;
+  } finally {
+    await db.close();
   }
-  if (name[0] === "@") {
-    const agent = await getAgentIdByNameOrCreate(name.slice(1), context);
-    const id = `@${agent.id.toString()}`;
-    return [id, agent];
-  }
-  return name;
+  // return `agent:${name}`;
 };
 
-const parsePostContent = async (
-  content: string,
-  mentions: (string | [string, Agent])[],
-): Promise<string> => {
-  const updatedContent = await replaceMentions(
-    content,
-    replaceAndAccumulate((x) => replaceAgents(x, content), mentions),
+const fix = (forwards: Forward[], doc: string) => {
+  const splitForwords = [];
+  for (const forward of forwards) {
+    const H = head(forward);
+    const T = tail(forward);
+    for (const h of H) {
+      const item = [h];
+      if (T) {
+        item.push(T);
+      }
+
+      splitForwords.push(item);
+    }
+  }
+  const orderedForwords = splitForwords.sort(([a], [b]) => {
+    if (typeof a !== "string" || typeof b !== "string") {
+      throw new Error("Expected string");
+    }
+    if (a.startsWith("#") && !b.startsWith("#")) {
+      return -1;
+    }
+    if (b.startsWith("#") && !a.startsWith("#")) {
+      return 1;
+    }
+    return docSort(doc)(a, b);
+  });
+
+  const mergedForwards = orderedForwords.reduce((p, current) => {
+    const previous = p.pop();
+    if (!previous) {
+      p.push(current);
+    } else {
+      if (previous[0] === current[0]) {
+        p.push(merge(previous as Forward, current, doc));
+      } else {
+        p.push(previous);
+        p.push(current);
+      }
+    }
+    return p;
+  }, []);
+
+  const sequential = [];
+  const simultaneous = [];
+
+  for (const [head, tail] of mergedForwards) {
+    const item = [head];
+    if (tail) {
+      item.push(tail);
+    }
+    if ((head as string).startsWith("#")) {
+      sequential.push(item);
+    } else {
+      simultaneous.push(item);
+    }
+  }
+  return [sequential, simultaneous];
+};
+
+export const processPost = async (
+  original: string,
+  startingForward: Forward = [],
+) => {
+  const match = new RegExp(MENTION_MATCH.source, MENTION_MATCH.flags);
+  const forwards: Forward[] = [];
+  if (startingForward) {
+    forwards.push(startingForward);
+  }
+  const dehydrated = await replaceAll(
+    original,
+    match,
+    async (exec?: RegExpExecArray) => {
+      const { padStart = "", padEnd = "" } = exec!.groups as {
+        padStart: string;
+        padEnd: string;
+      };
+      const forward: Exclude<Forward, string[]> = [];
+      let current: Exclude<Forward, string[]> = forward;
+      let prev: Exclude<Forward, string[]>;
+      const groups = exec![0].split("|");
+      let groupcount: number = groups.length;
+      for (const group of groups) {
+        for (let match of group.split(",")) {
+          match = match.trim();
+          if (match.startsWith("#")) {
+            current.push(match);
+          } else {
+            if (match.startsWith("@")) {
+              match = match.slice(1);
+            }
+            const id = await getAgentIdFromName(match, original);
+            current.push(id);
+          }
+        }
+        if (((groupcount -= 1) > 0)) {
+          prev = current;
+          current = [];
+          prev.push(current);
+        }
+      }
+      forwards.push(forward);
+      return `${padStart}${head(forward).join(",")}${padEnd}`;
+    },
   );
-  return updatedContent;
+
+  const [sequential, simultaneous] = fix(forwards, dehydrated);
+  return { original, dehydrated, sequential, simultaneous };
 };
 
-export default parsePostContent;
+export const blankPost = async (
+  original: string,
+  { mentions = false, hashtags = false }: {
+    mentions?: boolean | string[];
+    hashtags?: boolean | string[];
+  } = {},
+): Promise<string> => {
+  const match = new RegExp(MENTION_MATCH.source, MENTION_MATCH.flags);
+  const blanked = await replaceAll(
+    original,
+    match,
+    async (exec?: RegExpExecArray) => {
+      const { padStart = "", padEnd = "" } = exec!.groups as {
+        padStart: string;
+        padEnd: string;
+      };
+      const match = exec![0].trim();
+      if (match.startsWith("#") && hashtags) {
+        if (hashtags === true) {
+          return "";
+        }
+        if (hashtags.includes(match)) {
+          return "";
+        }
+      } else if (mentions) {
+        if (mentions === true) {
+          return "";
+        }
+        if (mentions.includes(match)) {
+          return "";
+        }
+      }
+      return `${padStart}${exec![0]}${padEnd}`;
+    },
+  );
+  return blanked;
+};
+
+export default processPost;
